:recycle: refactor: ERD 반영하여 domain 수정

+++ src/main/java/com/gdgoc/bondly/domain/QuizChoice.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 public class QuizChoice {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer choiceId;
+    private Long choiceId;
     @ManyToOne
     @JoinColumn(name = "question_id", nullable = false)
     private QuizQuestion question;
